+++ ProdApp/ProdManagerApp/WebApp/ApiControllers/ComponentsController.cs
@@ -95,6 +95,7 @@ namespace WebApp.ApiControllers
             }
             
             var component = await _repo.FirstOrDefaultAsync(id);
+            if (component == null) return NotFound();
             _repo.Remove(component);
             await _context.SaveChangesAsync();
